requirements.txt
ccxt==4.2.25
python-binance==1.0.19
websocket-client==1.7.0
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
coloredlogs==15.0.1
orjson==3.9.10